tests: version handling: move code
