+++ misc/python/materialize/version_list.py
@@ -13,7 +13,13 @@ from pathlib import Path
 
 import frontmatter
 
-from materialize import git
+from materialize import buildkite, git
+from materialize.docker import (
+    commit_to_image_tag,
+    image_of_commit_exists,
+    image_of_release_version_exists,
+    version_to_image_tag,
+)
 from materialize.git import get_version_tags
 from materialize.mz_version import MzVersion
 
@@ -34,6 +40,190 @@ INVALID_VERSIONS = {
     MzVersion.parse_mz("v0.57.5"),
     MzVersion.parse_mz("v0.57.6"),
 }
+
+
+"""
+Git revisions that are based on commits listed as keys require at least the version specified in the value.
+Note that specified versions do not necessarily need to be already published.
+Commits must be ordered descending by their date.
+"""
+MIN_ANCESTOR_MZ_VERSION_PER_COMMIT: dict[str, MzVersion] = {
+    # insert newer commits at the top
+    # PR#23421 (coord: smorgasbord of improvements for the crdb-backed timestamp oracle) introduces regressions against 0.78.13
+    "5179ebd39aea4867622357a832aaddcde951b411": MzVersion.parse_mz("v0.79.0")
+}
+
+
+def resolve_ancestor_image_tag() -> str:
+    image_tag, context = _resolve_ancestor_image_tag()
+    print(f"Using {image_tag} as image tag for ancestor (context: {context})")
+    return image_tag
+
+
+def _resolve_ancestor_image_tag() -> tuple[str, str]:
+    if buildkite.is_in_buildkite():
+        return _resolve_ancestor_image_tag_when_in_buildkite()
+    else:
+        return _resolve_ancestor_image_tag_when_running_locally()
+
+
+def _resolve_ancestor_image_tag_when_in_buildkite() -> tuple[str, str]:
+    if buildkite.is_in_pull_request():
+        # return the merge base
+        common_ancestor_commit = buildkite.get_merge_base()
+        if image_of_commit_exists(common_ancestor_commit):
+            return (
+                commit_to_image_tag(common_ancestor_commit),
+                "merge base of pull request",
+            )
+        else:
+            return (
+                version_to_image_tag(get_latest_published_version()),
+                "latest release because image of merge base of pull request not available",
+            )
+    elif git.is_on_release_version():
+        # return the previous release
+        tagged_release_version = git.get_tagged_release_version(version_type=MzVersion)
+        assert tagged_release_version is not None
+        previous_release_version = get_previous_published_version(
+            tagged_release_version
+        )
+        return (
+            version_to_image_tag(previous_release_version),
+            f"previous release because on release branch {tagged_release_version}",
+        )
+    else:
+        latest_published_version = get_latest_published_version()
+        override_commit = _get_override_commit_instead_of_version(
+            latest_published_version
+        )
+
+        if override_commit is not None:
+            # use the commit instead of the latest release
+            return (
+                commit_to_image_tag(override_commit),
+                f"commit override instead of latest release ({latest_published_version})",
+            )
+
+        # return the latest release
+        return (
+            version_to_image_tag(latest_published_version),
+            "latest release because not in a pull request and not on a release branch",
+        )
+
+
+def _resolve_ancestor_image_tag_when_running_locally() -> tuple[str, str]:
+    if git.is_on_release_version():
+        # return the previous release
+        tagged_release_version = git.get_tagged_release_version(version_type=MzVersion)
+        assert tagged_release_version is not None
+        previous_release_version = get_previous_published_version(
+            tagged_release_version
+        )
+        return (
+            version_to_image_tag(previous_release_version),
+            f"previous release because on local release branch {tagged_release_version}",
+        )
+    elif git.is_on_main_branch():
+        # return the latest release
+        latest_published_version = get_latest_published_version()
+        override_commit = _get_override_commit_instead_of_version(
+            latest_published_version
+        )
+
+        if override_commit is not None:
+            # use the commit instead of the latest release
+            return (
+                commit_to_image_tag(override_commit),
+                f"commit override instead of latest release ({latest_published_version})",
+            )
+
+        return (
+            version_to_image_tag(latest_published_version),
+            "latest release because on local main branch",
+        )
+    else:
+        # return the merge base
+        common_ancestor_commit = buildkite.get_merge_base()
+        if image_of_commit_exists(common_ancestor_commit):
+            return (
+                commit_to_image_tag(common_ancestor_commit),
+                "merge base of local non-main branch",
+            )
+        else:
+            return (
+                version_to_image_tag(get_latest_published_version()),
+                "latest release because image of merge base of local non-main branch not available",
+            )
+
+
+def _get_override_commit_instead_of_version(
+    latest_published_version: MzVersion,
+) -> str | None:
+    """
+    If a commit specifies a mz version as prerequisite (to avoid regressions) that is newer than the
+    provided latest version (i.e., prerequisite not satisfied by the latest version), then return
+    that commit's hash if the commit contained in the current state.
+    Otherwise, return none.
+    """
+    for (
+        commit_hash,
+        min_required_mz_version,
+    ) in MIN_ANCESTOR_MZ_VERSION_PER_COMMIT.items():
+        if latest_published_version >= min_required_mz_version:
+            continue
+
+        if git.contains_commit(commit_hash):
+            # commit would require at least min_required_mz_version
+            return commit_hash
+
+    return None
+
+
+def get_latest_published_version() -> MzVersion:
+    excluded_versions = set()
+
+    while True:
+        latest_published_version = git.get_latest_version(
+            version_type=MzVersion, excluded_versions=excluded_versions
+        )
+
+        if image_of_release_version_exists(latest_published_version):
+            return latest_published_version
+        else:
+            print(
+                f"Skipping version {latest_published_version} (image not found), trying earlier version"
+            )
+            excluded_versions.add(latest_published_version)
+
+
+def get_previous_published_version(release_version: MzVersion) -> MzVersion:
+    excluded_versions = set()
+
+    while True:
+        previous_published_version = get_previous_mz_version(
+            release_version, excluded_versions=excluded_versions
+        )
+
+        if image_of_release_version_exists(previous_published_version):
+            return previous_published_version
+        else:
+            print(f"Skipping version {previous_published_version} (image not found)")
+            excluded_versions.add(previous_published_version)
+
+
+def get_published_minor_mz_versions(limit: int | None = None) -> list[MzVersion]:
+    minor_mz_versions = get_all_minor_mz_versions()
+    versions = []
+
+    for v in minor_mz_versions:
+        if image_of_release_version_exists(v):
+            versions.append(v)
+
+        if limit is not None and len(versions) == limit:
+            break
+
+    return versions
 
 
 def get_all_mz_versions() -> list[MzVersion]:
