+++ test/scalability/mzcompose.py
@@ -15,7 +15,7 @@ import pandas as pd
 from jupyter_core.command import main as jupyter_core_command_main
 from matplotlib import pyplot as plt
 
-from materialize import buildkite, docker, git
+from materialize import buildkite, git
 from materialize.mzcompose.composition import Composition, WorkflowArgumentParser
 from materialize.mzcompose.services.materialized import Materialized
 from materialize.mzcompose.services.postgres import Postgres
@@ -45,6 +45,7 @@ from materialize.scalability.workload import Workload, WorkloadSelfTest
 from materialize.scalability.workloads import *  # noqa: F401 F403
 from materialize.scalability.workloads_test import *  # noqa: F401 F403
 from materialize.util import YesNoOnce, all_subclasses
+from materialize.version_list import resolve_ancestor_image_tag
 
 SERVICES = [
     Materialized(
@@ -251,7 +252,7 @@ def get_baseline_and_other_endpoints(
             )
         else:
             if target == "common-ancestor":
-                target = docker.resolve_ancestor_image_tag()
+                target = resolve_ancestor_image_tag()
             endpoint = MaterializeContainer(
                 composition=c,
                 specified_target=original_target,
